refactor: split API services into dedicated modules
