+++ src/pages/settings/ExternalServicesSettings.tsx
@@ -9,7 +9,7 @@ import { Globe, Plus, Settings, Trash2, Loader2, AlertCircle, RefreshCw } from '
 import { Alert, AlertDescription, AlertTitle } from '../../components/ui/alert';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '../../components/ui/dialog';
 import { useToast } from '../../components/ui/use-toast';
-import { api } from '../../services/api';
+import { api } from '../../services';
 
 interface ExternalService {
   id: string;
