+++ src/pages/settings/ReportTemplatesSettings.tsx
@@ -7,7 +7,7 @@ import { FileBarChart2, Plus, FileEdit, Trash2, Download, Loader2, AlertCircle, 
 import { Alert, AlertDescription, AlertTitle } from '../../components/ui/alert';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '../../components/ui/dialog';
 import { useToast } from '../../components/ui/use-toast';
-import { api } from '../../services/api';
+import { api } from '../../services';
 import { Input } from '../../components/ui/input';
 import { Label } from '../../components/ui/label';
 import { Textarea } from '../../components/ui/textarea';
