+++ src/pages/settings/ShippingMethodsSettings.tsx
@@ -9,7 +9,7 @@ import { Truck, Plus, Trash2, Package, Loader2, AlertCircle, X } from 'lucide-re
 import { Alert, AlertDescription, AlertTitle } from '../../components/ui/alert';
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from '../../components/ui/dialog';
 import { useToast } from '../../components/ui/use-toast';
-import { api } from '../../services/api';
+import { api } from '../../services';
 import { CountrySelector } from '../../components/selectors';
 import { useCountries } from '../../hooks/useReferenceData';
